some reorganisation
